Feature/7 요금제 생성 기능 예외 처리 및 리팩토링

+++ src/main/java/com/comprehensive/eureka/plan/dto/PlanDto.java
@@ -21,5 +21,5 @@ public class PlanDto {
     private Integer VoiceAllowance;
     private Integer additionalCallAllowance;
     private Integer monthlyFee;
-    private List<Integer> benefitIdList;
+    private List<Long> benefitIdList;
 }
